Add Constructor for DisabledPermissioner and implementation for DisabledUserManagement
